Add minimal complex support to the core library
Add two new types: complex float and complex double. The z-transform
provided at the creation of the filter can now be complex (complex input
is not yet supported). In that can, the type of the output is expected
to have the same precision than the one processed (input) but become
complex as well.