Fix initialization with rtf_init_filter()
The internal states of the filter are now set taking into account the DC
gain in the case of IIR. The past output was indeed previously
initialized to the same value as data argument. This is obviously
incorrect in the case of highpass for example. The initialization of
past output now takes into account the real steady state response of a
step.

It should be noted that normally, those modications should not change how
filters are initialized when they are FIR.

Change-Id: I47ce28fcb367c561868f1299e9e469567ab79543