rework rtfilter tests
Instead of testing rtfilter against a reference implementation
(here matlab's one) test the filters as themselves.
Eg. test a lowpass filter as "the object that let through low
frequencies and filters out high frequencies"

This means that we tests the *function* and not the *implementation*,
and this is the fundamental issue this work wishes to fix.

This drops the test dependency on matlab/octave, and use the check
framework to help writing the new tests.
This is the first commit of a serie that will test the rtfilter API;
it lays down the foundations and puts the new framework in place.

Test rtf_filter() on all data types, and separate the ones working on
float data type for later use with higher-level API functions.

Change-Id: Ie9e3ccaa869ea869b680c19ff6b1e70b29b3033b