global API change to use "int" instead of "unsigned int"
This is an API change that should not break the ABI.

The rationale is that "a number of ..." is expressed as an int in almost
all cases.
It has the nice side-effect of being shorter to write (and read).
It should make the API simpler and the code using the rtfilter API
easier to write and read.

API variables changed:
* ns, r (number of samples)
* nchann (number of channels)
* numlen, denum_len (number of coefficients)
* num_poles (number of poles)
* hlen (fir half-length)

Change-Id: I7f586777e249782bad47e6614763b88ca7074093