use PI as defined by math.h
keep defining-if-not M_PI on top of the file, though this should not be
necessary.

Only use the double version of PI in the whole file: the defined
variable was actually a float cast to double for the operation, then
cast to flow when writing the resulting value. This removes one of the
casts. Since the rest of the input (eg. fequencies) is given as doubles,
the other cast is mandatory.

remove some explicit casts (that were implicit) to improve lisibility.

Change-Id: I42a868614b6addbcb5af34c2e2ef6d8409b7fc3f