Add meson build system support
This is meant to coexist with the autotools build system.
This will allow a test period for meson as long as we want it to be.

Note: the minimum meson version required here is 0.47. This means that
debian 9 (current stable version) does not package a recent enough
version of meson for this to work (0.37.1). This has been tested with
debian 10 which packages meson 0.49.2.

Change-Id: Ia038b5f28aa2806c032145b8f12a619a625e37aa