api - remove use of anonymous union
Anonymous unions are not part of C99 standard, nor C++11 standard.

!! This is an API break !! (It does not break the ABI)
However, only one project that I could find uses that API for now.
I will update it as soon as this gets merged.

The change itself is pretty much transparent.
Replace:
    union {
        struct real_coeffs real_coeffs;
        struct complex_coeffs complex_coeffs;
    };
With this:
    union {
        struct real_coeffs real;
        struct complex_coeffs cmplx;
    } coeffs;

Use "cmplx" instead of "complex" to prevent a bad mix with the syntax
keyword.

Change-Id: Id0f67725fc5febf031992f889565e159ca11d3cd