Update patch set 1
Patch Set 1:

> I do not see why the build fails for win32-autotools but succeeds for the other ones.

The problem of win32-autotools is not your fault... It is an issue in new version of linux-doc... Upstream is aware of this but has not fixed it properly yet. I've just disabled doc generation in jenkins for win32-autotools.

Also, don't abandon a change just because there build that fail... Just fix it and push again. You should abandon a change only if you the concept of the change must be dropped. If it requires only correction, keep it.

RECHECK

Patch-set: 1